+++ lib/enums.dart
@@ -1,0 +1,12 @@
+/// This library combines enumerations used in Fetch API and is safe to use
+/// in non-JS environments.
+library fetch_api.enums;
+
+export 'src/request/request_cache.dart';
+export 'src/request/request_credentials.dart';
+export 'src/request/request_destination.dart';
+export 'src/request/request_duplex.dart';
+export 'src/request/request_mode.dart';
+export 'src/request/request_redirect.dart';
+export 'src/request/request_referrer_policy.dart';
+export 'src/response/response_type.dart';