+++ lib/src/dom_exception.dart
@@ -1,5 +1,0 @@
-import 'dart:js_interop';
-
-
-/// Placeholder for [DOMException].
-typedef DOMException = JSAny;